feat: [feature2] 리팩토링
