+++ src/Calculator.js
@@ -1,0 +1,17 @@
+export default class Calculator {
+  add(operand1, operand2) {
+    return operand1 + operand2;
+  }
+
+  subtarct(operand1, operand2) {
+    return operand1 - operand2;
+  }
+
+  multiply(operand1, operand2) {
+    return operand1 * operand2;
+  }
+
+  divide(operand1, operand2) {
+    return operand1 / operand2;
+  }
+}
