+++ src/constants.js
@@ -1,0 +1,5 @@
+// 연산 과정 및 오류 체크에 필요한 상수
+export const POS_INF = Infinity;
+export const NEG_INF = -Infinity;
+export const POS_ZERO = +0;
+export const NEG_ZERO = -0;
