--- src/index.js
@@ -1,17 +1,0 @@
-export default class Calculator {
-  add(operand1, operand2) {
-    return operand1 + operand2;
-  }
-
-  subtarct(operand1, operand2) {
-    return operand1 - operand2;
-  }
-
-  multiply(operand1, operand2) {
-    return operand1 * operand2;
-  }
-
-  divide(operand1, operand2) {
-    return operand1 / operand2;
-  }
-}
