+++ utils/constants.js
@@ -1,0 +1,4 @@
+export const POS_INF = Infinity;
+export const NEG_INF = -Infinity;
+export const POS_ZERO = +0;
+export const NEG_ZERO = -0;
